+++ messaging/src/main/java/de/fraunhofer/ids/messaging/endpoint/MessageController.java
@@ -37,6 +37,7 @@ import de.fraunhofer.ids.messaging.protocol.multipart.parser.MultipartDatapart;
 import de.fraunhofer.ids.messaging.util.IdsMessageUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -65,6 +66,12 @@ public class MessageController {
      * The infomodel serializer.
      */
     private final Serializer serializer;
+
+    /**
+     * Used to switch incoming infomodel version compatibility check off or on (default on).
+     */
+    @Value("#{new Boolean('${infomodel.compatibility.validation:true}')}")
+    private Boolean validateInfVer;
 
     /**
      * Constructor for the MessageController.
@@ -123,12 +130,12 @@ public class MessageController {
                 input = scanner.useDelimiter("\\A").next();
             }
 
-            if (!checkInboundVersion(input)) {
+            if (!validateInfomodelVersion(input)) {
                 return ResponseEntity
                         .status(HttpStatus.BAD_REQUEST)
                         .body(createDefaultErrorMessage(
-                             RejectionReason.VERSION_NOT_SUPPORTED,
-                             "Infomodel Version of incoming Message not supported!"));
+                                RejectionReason.VERSION_NOT_SUPPORTED,
+                                "Infomodel Version of incoming Message not supported!"));
             }
 
             // Deserialize JSON-LD headerPart to its RequestMessage.class
@@ -206,6 +213,29 @@ public class MessageController {
                              String.format(
                                  "Could not read incoming request! Error: %s", e.getMessage())));
         }
+    }
+
+    /**
+     * Decides whether to run the infomodel compatibility check and if so, whether the
+     * incoming message is compatible.
+     * @param input The received message.
+     * @return True if compatible or no validation to be performed, false otherwise.
+     * @throws IOException No model version information found in the header.
+     */
+    private boolean validateInfomodelVersion(final String input) throws IOException {
+        if (validateInfVer && !checkInboundVersion(input)) {
+            return false;
+        } else if (!validateInfVer) {
+            if (log.isDebugEnabled()) {
+                log.debug("Skipped validating infomodel compability!");
+            }
+        } else {
+            if (log.isInfoEnabled()) {
+                log.info("Successfully validated infomodel compability.");
+            }
+        }
+
+        return true;
     }
 
     /**